Remove CardStream legacy and replace it with CardStream Modern gateway
